+++ plugins/fs/src/watcher.rs
@@ -3,10 +3,10 @@
 // SPDX-License-Identifier: MIT
 
 use notify::{Config, Event, RecommendedWatcher, RecursiveMode, Watcher};
-use notify_debouncer_mini::{new_debouncer, DebounceEventResult, Debouncer};
+use notify_debouncer_full::{new_debouncer, DebounceEventResult, Debouncer, FileIdMap};
 use serde::Deserialize;
 use tauri::{
-    ipc::Channel,
+    ipc::{Channel, CommandScope, GlobalScope},
     path::{BaseDirectory, SafePathBuf},
     AppHandle, Manager, Resource, ResourceId, Runtime,
 };
@@ -21,7 +21,10 @@ use std::{
     time::Duration,
 };
 
-use crate::commands::{resolve_path, CommandResult};
+use crate::{
+    commands::{resolve_path, CommandResult},
+    scope::Entry,
+};
 
 struct InnerWatcher {
     pub kind: WatcherKind,
@@ -43,7 +46,7 @@ impl WatcherResource {
 impl Resource for WatcherResource {}
 
 enum WatcherKind {
-    Debouncer(Debouncer<RecommendedWatcher>),
+    Debouncer(Debouncer<RecommendedWatcher, FileIdMap>),
     Watcher(RecommendedWatcher),
 }
 
@@ -60,10 +63,10 @@ fn watch_raw(on_event: Channel, rx: Receiver<notify::Result<Event>>) {
 
 fn watch_debounced(on_event: Channel, rx: Receiver<DebounceEventResult>) {
     spawn(move || {
-        while let Ok(event) = rx.recv() {
-            if let Ok(event) = event {
+        while let Ok(Ok(events)) = rx.recv() {
+            for event in events {
                 // TODO: Should errors be emitted too?
-                let _ = on_event.send(&event);
+                let _ = on_event.send(&event.event);
             }
         }
     });
@@ -83,10 +86,18 @@ pub async fn watch<R: Runtime>(
     paths: Vec<SafePathBuf>,
     options: WatchOptions,
     on_event: Channel,
+    global_scope: GlobalScope<'_, Entry>,
+    command_scope: CommandScope<'_, Entry>,
 ) -> CommandResult<ResourceId> {
     let mut resolved_paths = Vec::with_capacity(paths.capacity());
     for path in paths {
-        resolved_paths.push(resolve_path(&app, path, options.dir)?);
+        resolved_paths.push(resolve_path(
+            &app,
+            &global_scope,
+            &command_scope,
+            path,
+            options.dir,
+        )?);
     }
 
     let mode = if options.recursive {
@@ -97,10 +108,9 @@ pub async fn watch<R: Runtime>(
 
     let kind = if let Some(delay) = options.delay_ms {
         let (tx, rx) = channel();
-        let mut debouncer = new_debouncer(Duration::from_millis(delay), tx)?;
-        let watcher = debouncer.watcher();
+        let mut debouncer = new_debouncer(Duration::from_millis(delay), None, tx)?;
         for path in &resolved_paths {
-            watcher.watch(path.as_ref(), mode)?;
+            debouncer.watcher().watch(path.as_ref(), mode)?;
         }
         watch_debounced(on_event, rx);
         WatcherKind::Debouncer(debouncer)
@@ -130,14 +140,14 @@ pub async fn unwatch<R: Runtime>(app: AppHandle<R>, rid: ResourceId) -> CommandR
                 for path in &watcher.paths {
                     debouncer.watcher().unwatch(path.as_ref()).map_err(|e| {
                         format!("failed to unwatch path: {} with error: {e}", path.display())
-                    })?
+                    })?;
                 }
             }
             WatcherKind::Watcher(ref mut w) => {
                 for path in &watcher.paths {
                     w.unwatch(path.as_ref()).map_err(|e| {
                         format!("failed to unwatch path: {} with error: {e}", path.display())
-                    })?
+                    })?;
                 }
             }
         }