+++ plugins/notification/android/src/main/java/Notification.kt
@@ -80,8 +80,6 @@ class Notification {
     return null
   }
 
-  val isScheduled = schedule != null
-
   companion object {
     fun buildNotificationPendingList(notifications: List<Notification>): List<PendingNotification> {
       val pendingNotifications = mutableListOf<PendingNotification>()