+++ plugins/updater/src/config.rs
@@ -2,26 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
+use std::{ffi::OsString, fmt::Display};
+
 use serde::{Deserialize, Deserializer};
 use url::Url;
 
-/// Updater configuration.
-#[derive(Debug, Clone, Deserialize, Default)]
-pub struct Config {
-    #[serde(default)]
-    pub endpoints: Vec<UpdaterEndpoint>,
-    /// Signature public key.
-    pub pubkey: String,
-    /// Additional arguments given to the NSIS or WiX installer.
-    #[serde(default, alias = "installer-args")]
-    pub installer_args: Vec<String>,
-    /// The Windows configuration for the updater.
-    #[serde(default)]
-    pub windows: UpdaterWindowsConfig,
-}
-
 /// Install modes for the Windows update.
-#[derive(Debug, PartialEq, Eq, Clone)]
+#[derive(Debug, PartialEq, Eq, Clone, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub enum WindowsUpdateInstallMode {
     /// Specifies there's a basic UI during the installation process, including a final dialog box at the end.
     BasicUi,
@@ -30,43 +18,9 @@ pub enum WindowsUpdateInstallMode {
     Quiet,
     /// Specifies unattended mode, which means the installation only shows a progress bar.
     Passive,
-    // to add more modes, we need to check if the updater relaunch makes sense
-    // i.e. for a full UI mode, the user can also mark the installer to start the app
-}
-
-impl Default for WindowsUpdateInstallMode {
-    fn default() -> Self {
-        Self::Passive
-    }
-}
-
-impl<'de> Deserialize<'de> for WindowsUpdateInstallMode {
-    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        match s.to_lowercase().as_str() {
-            "basicui" => Ok(Self::BasicUi),
-            "quiet" => Ok(Self::Quiet),
-            "passive" => Ok(Self::Passive),
-            _ => Err(serde::de::Error::custom(format!(
-                "unknown update install mode '{s}'"
-            ))),
-        }
-    }
 }
 
 impl WindowsUpdateInstallMode {
-    /// Returns the associated nsis arguments.
-    pub fn nsis_args(&self) -> &'static [&'static str] {
-        match self {
-            Self::Passive => &["/P", "/R"],
-            Self::Quiet => &["/S", "/R"],
-            _ => &[],
-        }
-    }
-
     /// Returns the associated `msiexec.exe` arguments.
     pub fn msiexec_args(&self) -> &'static [&'static str] {
         match self {
@@ -75,13 +29,59 @@ impl WindowsUpdateInstallMode {
             Self::Passive => &["/passive"],
         }
     }
+
+    /// Returns the associated nsis arguments.
+    pub fn nsis_args(&self) -> &'static [&'static str] {
+        match self {
+            Self::Passive => &["/P", "/R"],
+            Self::Quiet => &["/S", "/R"],
+            _ => &[],
+        }
+    }
 }
 
-#[derive(Debug, Clone, Default, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct UpdaterWindowsConfig {
+impl Display for WindowsUpdateInstallMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Self::BasicUi => "basicUI",
+                Self::Quiet => "quiet",
+                Self::Passive => "passive",
+            }
+        )
+    }
+}
+
+impl Default for WindowsUpdateInstallMode {
+    fn default() -> Self {
+        Self::Passive
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct WindowsConfig {
+    /// Additional arguments given to the NSIS or WiX installer.
+    #[serde(default, alias = "installer-args")]
+    pub installer_args: Vec<OsString>,
+    /// Updating mode, see [`WindowsUpdateInstallMode`] for more info.
     #[serde(default, alias = "install-mode")]
     pub install_mode: WindowsUpdateInstallMode,
+}
+
+/// Updater configuration.
+#[derive(Debug, Clone, Deserialize, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct Config {
+    /// Updater endpoints.
+    #[serde(default)]
+    pub endpoints: Vec<UpdaterEndpoint>,
+    /// Signature public key.
+    pub pubkey: String,
+    /// The Windows configuration for the updater.
+    pub windows: Option<WindowsConfig>,
 }
 
 /// A URL to an updater server.